Simplified default interface for JsonSchemaGenerator when loading a model into it
