+++ CsCore/CsCoreUnity/Plugins/CsCoreUnity/com/csutil/ui/jsonschema/JsonModelExtensions.cs
@@ -11,14 +11,21 @@ namespace com.csutil.ui.mtvmtv {
     /// <summary> Meothds that help connecting a JSON model instance to their schema views during runtime </summary>
     public static class JsonModelExtensions {
 
-        /// <summary> 
-        /// Converts the passed model to JSON, lets the user edit it and returned a parsed back clone with all changes 
-        /// made by the user, so that this new state can be stored or the changed fields can be calculated via MergeJson.GetDiff()
-        /// </summary>
+        /// <summary> Converts the passed model to JSON, lets the user edit it and returned a 
+        /// parsed back clone with all changes made by the user, so that this new state can be 
+        /// stored or the changed fields can be calculated via MergeJson.GetDiff() </summary>
         /// <param name="model"> The model that should be shown in the UI (has to fit the loaded view model UI) </param>
-        /// <param name="userSavedChanges"> 
-        /// A task that should be set to completed once the user is finished with the UI, e.g. when he presses the save button 
-        /// </param>
+        /// <returns> The modified model after the passed userSavedChanges-Task is completed </returns>
+        public static async Task<T> LoadViaJsonIntoView<T>(this Presenter<JObject> self, T model) {
+            return await LoadViaJsonIntoView(self, model, JsonSchemaPresenter.ChangesSavedViaConfirmButton(self.targetView));
+        }
+
+        /// <summary> Converts the passed model to JSON, lets the user edit it and returned a 
+        /// parsed back clone with all changes made by the user, so that this new state can be 
+        /// stored or the changed fields can be calculated via MergeJson.GetDiff() </summary>
+        /// <param name="model"> The model that should be shown in the UI (has to fit the loaded view model UI) </param>
+        /// <param name="userSavedChanges">  A task that should be set to completed once the user is finished with 
+        /// the UI, e.g. when he presses the save button </param>
         /// <returns> The modified model after the passed userSavedChanges-Task is completed </returns>
         public static async Task<T> LoadViaJsonIntoView<T>(this Presenter<JObject> self, T model, Task userSavedChanges) {
             JObject json = JObject.Parse(JsonWriter.GetWriter().Write(model));
