--- CsCore/CsCoreUnity/Plugins/CsCoreUnity/com/csutil/ui/jsonschema/VmtvContainerHelper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 934fb4ea53b964847b0ffc4e23d20a0e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
